% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_ci.R
\name{ar_ci}
\alias{ar_ci}
\title{Calculate Anderson-Rubin Confidence Intervals for fixest IV objects
(Compatible with Fixed Effects)}
\usage{
ar_ci(model, param, level = 0.95, grid_range = NULL, grid_length = 1000)
}
\arguments{
\item{model}{A 'fixest' object resulting from an IV estimation.}

\item{param}{A character string specifying the name of the endogenous variable.
If the model has only one, this can be omitted.}

\item{level}{The confidence level required (e.g., 0.95 for a 95\% CI).}

\item{grid_range}{A numeric vector of length 2, \code{c(min, max)}, specifying the
range of coefficient values to test. If NULL, a default is chosen.}

\item{grid_length}{An integer specifying the number of points in the grid to test.}
}
\value{
Returns an object of class 'ar_ci' containing the confidence interval,
the grid of tested values, and their corresponding p-values.
}
\description{
This function calculates Anderson-Rubin (AR) confidence intervals for a specified
endogenous variable from an IV regression estimated with \code{fixest::feols}.
It is robust to weak instruments and compatible with models containing fixed effects.
}
