% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar_ci.R
\name{get_ar_ci}
\alias{get_ar_ci}
\title{For a regression table, get Anderson-Rubin Confidence Intervals for fixest IV objects
(Compatible with Fixed Effects)}
\usage{
get_ar_ci(model)
}
\arguments{
\item{model}{A 'fixest' object resulting from an IV estimation.}
}
\value{
Returns a string representing the Anderson-Rubin confidence interval.
}
\description{
This function displays, after calculating, Anderson-Rubin (AR) confidence intervals for a specified
endogenous variable from an IV regression estimated with \code{fixest::feols}.
It is robust to weak instruments and compatible with models containing fixed effects.
}
