% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ar_ci.R
\name{plot.ar_ci}
\alias{plot.ar_ci}
\title{Plot method for ar_ci objects}
\usage{
\method{plot}{ar_ci}(x, ...)
}
\arguments{
\item{x}{An object of class 'ar_ci' returned by the \code{ar_ci} function.}

\item{...}{Additional arguments passed to the plot function.}
}
\description{
Plot method for ar_ci objects
}
