% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ar_ci.R
\name{print.ar_ci}
\alias{print.ar_ci}
\title{Print method for ar_ci objects}
\usage{
\method{print}{ar_ci}(x, ...)
}
\arguments{
\item{x}{An object of class 'ar_ci' returned by the \code{ar_ci} function.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for ar_ci objects
}
